+++ extensions/auth/ranger/impl/src/test/java/org/apache/polaris/extension/auth/ranger/RangerTestUtils.java
@@ -1,0 +1,100 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.polaris.extension.auth.ranger;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
+import org.apache.polaris.core.config.PolarisConfiguration;
+import org.apache.polaris.core.config.RealmConfig;
+import org.apache.polaris.core.context.RealmContext;
+import org.apache.polaris.core.entity.CatalogEntity;
+
+public class RangerTestUtils {
+
+  private static final String REALM_CONTEXT_NAME = "POLARIS";
+
+  public static RangerPolarisAuthorizerConfig createConfig() {
+    Map<String, String> properties = new HashMap<>();
+
+    properties.put(
+        "authz.default.policy.source.impl",
+        "org.apache.ranger.admin.client.EmbeddedResourcePolicySource");
+    properties.put("authz.default.policy.source.embedded_resource.path", "/authz_tests");
+
+    return createConfig("dev_polaris", properties);
+  }
+
+  public static RangerPolarisAuthorizerConfig createConfig(
+      String serviceName, Map<String, String> properties) {
+    return new RangerPolarisAuthorizerConfig() {
+      @Override
+      public Optional<String> serviceName() {
+        return Optional.ofNullable(serviceName);
+      }
+
+      @Override
+      public Map<String, String> properties() {
+        return properties;
+      }
+    };
+  }
+
+  public static RealmConfig createRealmConfig() {
+    return new RealmConfig() {
+      @SuppressWarnings({"removal"})
+      @Override
+      public <T> T getConfig(String configName) {
+        return null;
+      }
+
+      @SuppressWarnings({"removal"})
+      @Override
+      public <T> T getConfig(String configName, T defaultValue) {
+        return null;
+      }
+
+      @Override
+      public <T> T getConfig(PolarisConfiguration<T> config) {
+        return config.defaultValue();
+      }
+
+      @Override
+      public <T> T getConfig(PolarisConfiguration<T> config, CatalogEntity catalogEntity) {
+        return config.defaultValue();
+      }
+
+      @Override
+      public <T> T getConfig(
+          PolarisConfiguration<T> config, Map<String, String> catalogProperties) {
+        return config.defaultValue();
+      }
+    };
+  }
+
+  public static RealmContext createRealmContext() {
+    return new RealmContext() {
+      @Override
+      public String getRealmIdentifier() {
+        return REALM_CONTEXT_NAME;
+      }
+    };
+  }
+}